added support for multiple texture slots to blender addon
Importer and exporter now support multiple texture slots. The first slot in the
material is the diffuse map, second and third slot are specular and normal map.
Other slots are ignored. All textures share the same coordinates.
The option to rotate the mesh is now enabled by default in importer and
exporter.