added new option 'fully opaque' to blender addon
This new option is a mesh property located in the G3D properties panel. If
enabled it sets the opacity of the mesh to 1.0, ignoring what's set as alpha in
materials which is normally used for the opacity. This allows the modder to set
alpha in materials to 0 and see the alpha effects of the texture in the 3D view
of blender.