blender addon: fixed vertex duplication and animation import
When a vertex got duplicated because it has a different texcoord, it was not
checked if this vertex-texcoord-combination was created before. So we get
unnecessary duplicates with exact same vertex-texcoord-pair (taking part in
different faces). This is now hopefully checked correctly.

When importing the animation of a model, the last shapekey wasn't activated in
Blender for the last frame. So, the last frame was the same as the first
because the first frame is the normal model with no shapekey activated. This
should be fixed now too.